update quota function
